server: return 401 on login if X-Requested-With header is send
This should detect AJAX requests and return error instead of redirect if
user is not logged in. This should simplify handling in the fro

The X-Requested-With header is not used by browsers but is a standard
header specified by background (AJAX) fetch. The value according to some
sources can be only 'XMLHttpRequest' but as it seems there can be
anything according to some of the tutorials I saw. It is also not the
target use of X-Requested-With header as it is rather intended for
security and recognition reasons. The only one who can set it is script
running on the same domain as an endpoint and thus it can be used to
verify requester. In the end, we only technically sniff the traffic so
we are running on the same domain and the only issue there could be if
someone tried to use turris-auth protected API between routers for
example.