Add translations
This adds translations for this project.

The Makefile here is kind of overkill but it wasn't easy to configure in
a good generic way babel to extract messages only in submodule. This
makes it possible. We use babel here but how we extract messages is
defined by Makefile and not by setup.py. The install integration is also
simply call to make as part of install process.